+++ structurizr-core/src/com/structurizr/view/Branding.java
@@ -11,12 +11,6 @@ public final class Branding {
 
     private Font font;
 
-    private ColorPair color1;
-    private ColorPair color2;
-    private ColorPair color3;
-    private ColorPair color4;
-    private ColorPair color5;
-
     Branding() {
     }
 
@@ -50,46 +44,6 @@ public final class Branding {
      */
     public void setFont(Font font) {
         this.font = font;
-    }
-
-    public ColorPair getColor1() {
-        return color1;
-    }
-
-    public void setColor1(ColorPair color1) {
-        this.color1 = color1;
-    }
-
-    public ColorPair getColor2() {
-        return color2;
-    }
-
-    public void setColor2(ColorPair color2) {
-        this.color2 = color2;
-    }
-
-    public ColorPair getColor3() {
-        return color3;
-    }
-
-    public void setColor3(ColorPair color3) {
-        this.color3 = color3;
-    }
-
-    public ColorPair getColor4() {
-        return color4;
-    }
-
-    public void setColor4(ColorPair color4) {
-        this.color4 = color4;
-    }
-
-    public ColorPair getColor5() {
-        return color5;
-    }
-
-    public void setColor5(ColorPair color5) {
-        this.color5 = color5;
     }
 
 }